+++ tests/unit/test_admin.py
@@ -8,7 +8,18 @@ from django.http import HttpRequest
 from django.test import TestCase
 from django.test.client import Client
 from django.urls import reverse
-from student.models import AnonymousUserId
+# Juniper compatibility: in Juniper, importing from
+# common.djangoapps.student.models raises
+# "RuntimeError: Model class common.djangoapps.student.models.AnonymousUserId
+# doesn't declare an explicit app_label and isn't in an application in
+# INSTALLED_APPS". So as not to rely on admins modifying INSTALLED_APPS,
+# fall back to using the old import syntax.
+#
+# This can be removed once we give up Juniper compatibility.
+try:
+    from common.djangoapps.student.models import AnonymousUserId
+except RuntimeError:
+    from student.models import AnonymousUserId
 from hastexo.admin import StackAdmin
 from hastexo.models import Stack