+++ Changelog.md
@@ -1,5 +1,7 @@
 Unreleased
 --------------------------
+* [Enhancement] Use full names for common.djangoapps imports from
+  edx-platform.
 * [Enhancement] Update requirements for Open EdX Koa release.
 
 Version 5.0.3 (2021-03-22)